pids are now uint for all the os's

+++ list_libs_linux.go
@@ -14,7 +14,7 @@ import (
 // HasLibrary checks if a process with a given pid has a certain library.
 // To do this we use the /proc/<pid>/maps to know which files are mapped.
 // the file format is described in `man proc`
-func HasLibrary(pid int, r *regexp.Regexp) (bool, error) {
+func HasLibrary(pid uint, r *regexp.Regexp) (bool, error) {
 	path := filepath.Join("/proc", strconv.Itoa(pid), "maps")
 	f, err := os.Open(path)
 	if err != nil {
@@ -49,7 +49,7 @@ func HasLibrary(pid int, r *regexp.Regexp) (bool, error) {
 
 // FindProcWithLib returns a list of process ids that have the given library loaded in memory.
 // It works looking at all the pids listed in /proc folder, and for each of them, checking its maps file.
-func FindProcWithLib(r *regexp.Regexp) ([]int, error) {
+func FindProcWithLib(r *regexp.Regexp) ([]uint, error) {
 	files, _ := ioutil.ReadDir("/proc/")
 	res := make([]int, 0)
 
@@ -59,12 +59,12 @@ func FindProcWithLib(r *regexp.Regexp) ([]int, error) {
 			continue
 		}
 
-		if has, err := HasLibrary(pid, r); err != nil {
+		if has, err := HasLibrary(uint(pid), r); err != nil {
 			//TODO(mvanotti): How should we report errors for multiple files? maybe a map[filepath]error ?
 			log.Println(err)
 			continue
 		} else if has {
-			res = append(res, pid)
+			res = append(res, uint(pid))
 		}
 	}
